backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
gunicorn==22.0.0
psycopg2==2.9.9
sqlparse==0.5.0
whitenoise==6.7.0
djangorestframework>=3.15,<4.0
requests>=2.31,<3.0